Fix type hint of `key` on `Visitable.__class_getitem__()`
`Visitable.__class_getitem__()` is defined to allow parameterized generics
in older versions of Python. (Issue #6759)
Previously, arg `key` of that function was type-hinted as `str`.
This caused some type checkers, including PyCharm, to emit a warning
if you tried to parameterize a function with the actual class it would return
instead of a string reference to it.

Now the type hint has been changed to `Any`, fixing this issue.

The new type of `Any` was selected because:
- `Union[str, type]` would not allow parameterizing
  with an optional type or a union of types
- `Union[str, type, types.UnionType]` would not work
  in versions of Python older than 3.10 (when `UnionType` was added)

The definition of `__class_getitem__()` in `registry.generate_base()`
has also been updated similarly.

Fixes: #9878