Support dataclass default with init=False
Fixed an issue where generating dataclasses fields that specified a
default`` value and set ``init=False`` would not work.
The dataclasses behavior in this case is to set the default
value on the class, that's not compatible with the descriptors used
by SQLAlchemy. To support this case the default is transformed to
a ``default_factory`` when generating the dataclass.

Fixes: #9879
Change-Id: I5151d388232eacd506a100ba18ce26970bf83cf3