indicate pymssql as only binary
a new release of pymssql is out and they don't have wheel files,
so builds are failing.

we don't have deps set up to build pymssql, so hack into tox.ini
to force pip to only use a wheel.

additionally, add "py3" as a recognized target where we are referring
to specific py3 versions.

Change-Id: Id876fe902d4d7cae964a241c9fd7b678645b0ea5