Oracle default arraysize is now set by the driver
Changed the default arraysize of the Oracle dialects so that the value set
by the driver is used, that is 100 at the time of writing for both
cx_oracle and oracledb. Previously the value was set to 50 by default.

Fixes: #10877
Change-Id: Ie4c53f42437d3d7dbbad36398d7883472577f367
(cherry picked from commit c8214ad4389284dc9508e49aeca701e5bf164454)