restore uuid_data_type as closed at top level
the supports_native_uuid attribute does NOT indicate
the UUID datatype being present, only that Uuid(native_uuid=True)
would be able to produce something. On SQL Server it produces
UNIQUEIDENTIFIER.  The current use for this requirement is
that of testing the uppercase UUID type that has to match
that exactly.

Change-Id: I050e5d1889f804ee3763b84828f2bd6a47dd265e